requirements.txt
aiogram==3.21.0
langchain==0.3.26
langchain-openai==0.3.27
openai==1.97.0
peewee==3.17.6
python-dotenv==1.0.1
